+++ .gitignore
@@ -149,3 +149,6 @@ configs/local/default.yaml
 /data/
 /logs/
 .env
+
+# Aim logging
+.aim
