+++ configs/experiment/example.yaml
@@ -36,3 +36,5 @@ logger:
   wandb:
     tags: ${tags}
     group: "mnist"
+  aim:
+    experiment: "mnist"
