+++ environment.yaml
@@ -21,9 +21,9 @@ channels:
 # compatibility is usually guaranteed
 
 dependencies:
-  - pytorch>=1.10
-  - torchvision>=0.11
-  - pytorch-lightning=1.*
+  - pytorch=2.*
+  - torchvision=0.*
+  - pytorch-lightning=2.*
   - torchmetrics=0.*
   - hydra-core=1.*
   - rich=13.*
