+++ requirements.txt
@@ -1,8 +1,8 @@
 # --------- pytorch --------- #
-torch>=1.10.0
-torchvision>=0.11.0
-pytorch-lightning==1.9.1
-torchmetrics==0.11.0
+torch>=2.0.0
+torchvision>=0.15.0
+lightning>=2.0.0
+torchmetrics>=0.11.4
 
 # --------- hydra --------- #
 hydra-core==1.3.2
@@ -14,6 +14,7 @@ hydra-optuna-sweeper==1.2.0
 # neptune-client
 # mlflow
 # comet-ml
+# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550
 
 # --------- others --------- #
 pyrootutils     # standardizing the project root setup
